Address David's comments.
